backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cabana-pizza-backend"
version = "0.1.0"
description = "Backend API for Cabana da Pizza"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
  "alembic>=1.16,<2.0",
  "fastapi>=0.115,<1.0",
  "pwdlib[argon2]>=0.2,<1.0",
  "PyJWT>=2.9,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "pydantic-settings>=2.10,<3.0",
  "python-multipart>=0.0.9,<1.0",
  "python-dotenv>=1.1,<2.0",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.34,<1.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
